trash commit

+++ 3d-1/src/main/java/com/nocmok/opengl/_3d/_3d/Screen.java
@@ -5,7 +5,7 @@ public interface Screen {
     // polygon[i][1] - y - координата i-й точки
     void drawPolygon(int[][] polygon);
     // Преобразует нормализованные координаты точки в диапазоне [0,1] в координаты этого экрана
-    int[] transform(float[] point, int[] screenPoint);
+    int[] transform(double[] point, int[] screenPoint);
     int getXSize();
     int getYSize();
 }
